Some changes on MongoWP.java
MongoWP.java is a compendium of constants. Several of its contants were
not related with MongoWP but with specific version of a Mongo server. A
lot of these fake constants have been removed. Besides,
MongoWP.ErrorCode inner class has been move to a external class.

This commit breaks binary compatibility!