Several changes to improve memory footprint of big requests
Some changes have been done to improve memory footprint of big requests.
The most important change is the operation insert, that does not
retrieve from Netty buffers all the documents that have to be inserted
as soon as it can. Instead, this command contains now a Guava 
FluentIterator (something similar to JDK8 Streams) which can be used
later read documents in a lazy way. This is very helpful, for example,
on ToroDB, where several transformations are done to these documents.
With the new API, ToroDB can lazily do the conversions on iterables and,
at the end of the chain, materialize the result. The generated memory is
the same, but all the intermediate objects can be almost instantly
garbage collected if the memory preasure is high.

The bad news is that this new system requires to maintain a reference to
the Netty buffer. That reference must be release once the buffer is not
needed. In that case, it is released when the request has been executed.
To do so, requests must implement AutoCloseable, which implies to modify
all requests. At the same time, some improvements on the separation of
concerns have been done to request decoders and some new MongoExceptions
have been created to improve the returned information in case of errors
on MongoWP