Removed the install section, as it could produce problems
The install section said something like that:
install:
  - mvn dependency:go-offline -q

The idea behind that is that dependencies can be downloaded before the
script is executed, so the script is easier to read. But... that
pluggin also tries to resolve the dependencies to other project defined
on the same multiproject tree (aka "sibling projects"). This can work
meanwhile Nexus contains the previous snapshots, but not when a new
release is done.